Add many features. * common-src/amfeatures.c (am_set_default_feature_set): New function   to set the default features set. * common-src/amfeatures.h (am_set_default_feature_set): Prototype. * common-src/amfeatures.h: Rename amanda_feature_auth_keyword to   fe_options_auth. * client-src/client_util.c: Check features. * server-src/amcheck.c: Ditto. * server-src/diskfile.c: Ditto. * server-src/driverio.c: Ditto. * server-src/planner.c: Ditto. * server-src/diskfile.h (optionstr): Change prototype. * client-src/sendsize.c: Fix compiler warning.

+++ client-src/sendsize.c
@@ -24,7 +24,7 @@
  * file named AUTHORS, in the root directory of this distribution.
  */
 /* 
- * $Id: sendsize.c,v 1.130 2002/04/17 20:06:10 martinea Exp $
+ * $Id: sendsize.c,v 1.131 2002/04/19 14:24:12 martinea Exp $
  *
  * send estimated backup sizes using dump
  */
@@ -379,7 +379,7 @@ char **argv;
 	    }
 	    if(est == NULL) {
 		dbprintf(("%s: unexpected child %ld\n",
-			  debug_prefix_time(NULL), child_pid));
+			  debug_prefix_time(NULL), (long)child_pid));
 	    } else {
 		est->done = 1;
 		dumpsrunning--;
